+++ ecs/event/event_test.go
@@ -1,6 +1,7 @@
 package event_test
 
 import (
+	"fmt"
 	"testing"
 
 	"github.com/mlange-42/arche/ecs/event"
@@ -15,4 +16,25 @@ func TestSubscriptions(t *testing.T) {
 
 	assert.True(t, m1.ContainsAny(event.ComponentAdded|event.TargetChanged))
 	assert.False(t, m1.Contains(event.ComponentAdded|event.RelationChanged))
+}
+
+func ExampleSubscription() {
+	mask := event.EntityCreated | event.EntityRemoved
+
+	fmt.Printf("%08b contains\n%08b -> %t\n\n", mask, event.EntityRemoved, mask.Contains(event.EntityRemoved))
+	fmt.Printf("%08b contains\n%08b -> %t\n\n", mask, event.ComponentAdded, mask.Contains(event.ComponentAdded))
+
+	fmt.Printf("%08b contains any\n%08b -> %t\n\n", mask, event.EntityRemoved|event.ComponentAdded, mask.ContainsAny(event.EntityRemoved|event.ComponentAdded))
+	fmt.Printf("%08b contains any\n%08b -> %t\n\n", mask, event.ComponentAdded|event.ComponentRemoved, mask.ContainsAny(event.ComponentAdded|event.ComponentRemoved))
+	// Output: 00000011 contains
+	// 00000010 -> true
+	//
+	// 00000011 contains
+	// 00000100 -> false
+	//
+	// 00000011 contains any
+	// 00000110 -> true
+	//
+	// 00000011 contains any
+	// 00001100 -> false
 }