+++ ecs/world_examples_test.go
@@ -276,11 +276,11 @@ func ExampleWorld_Batch() {
 func ExampleWorld_SetListener() {
 	world := ecs.NewWorld()
 
-	listener := NewTestListener(
-		func(evt ecs.EntityEvent) {
+	listener := TestListener{
+		Callback: func(world *ecs.World, evt ecs.EntityEvent) {
 			fmt.Println(evt)
 		},
-	)
+	}
 	world.SetListener(&listener)
 
 	world.NewEntity()
@@ -296,22 +296,23 @@ func ExampleWorld_Stats() {
 
 // TestListener for all [EntityEvent]s.
 type TestListener struct {
-	Callback func(e ecs.EntityEvent)
+	Callback func(world *ecs.World, evt ecs.EntityEvent)
 }
 
-// NewTestListener creates a new [TestListener] that subscribes to all event types.
-func NewTestListener(callback func(e ecs.EntityEvent)) TestListener {
-	return TestListener{
-		Callback: callback,
-	}
+// Notify the listener.
+// Calls the Callback.
+func (l *TestListener) Notify(world *ecs.World, evt ecs.EntityEvent) {
+	l.Callback(world, evt)
 }
 
-// Notify the listener
-func (l *TestListener) Notify(e ecs.EntityEvent) {
-	l.Callback(e)
-}
-
-// Subscriptions of the listener
+// Subscriptions of the listener.
+// Subscribes to all events.
 func (l *TestListener) Subscriptions() event.Subscription {
 	return event.All
+}
+
+// Components the listener subscribes to.
+// Subscribes to all components.
+func (l *TestListener) Components() *ecs.Mask {
+	return nil
 }