+++ examples/change_listener/main.go
@@ -27,30 +27,15 @@ type Rotation struct {
 	A float64
 }
 
-// TestListener type
-type TestListener struct {
-	World *ecs.World
-}
-
-// Notify is called on entity changes.
-func (l *TestListener) Notify(evt ecs.EntityEvent) {
-	// Just prints out what the event is about.
-	// This could also be a method of a type that manages events.
-	// Could use e.g. filters to distribute events to interested/registered systems.
-	if evt.Contains(event.EntityCreated) {
-		fmt.Printf("Entity added, has components %v\n", l.World.Ids(evt.Entity))
-	} else if evt.Contains(event.EntityRemoved) {
-		fmt.Printf("Entity removed, had components %v\n", l.World.Ids(evt.Entity))
-	} else {
-		fmt.Printf("Entity changed, has components %v\n", l.World.Ids(evt.Entity))
-	}
-}
-
 func main() {
 	// Create a World.
 	world := ecs.NewWorld()
-	ls := TestListener{World: &world}
-	wrapper := listener.NewCallback(event.Entities|event.Components, ls.Notify)
+	listener := listener.NewCallback(
+		func(world *ecs.World, evt ecs.EntityEvent) {
+			fmt.Printf("Events: %08b Added: %v Removed: %v\n", evt.EventTypes, evt.Added, evt.Removed)
+		},
+		event.Entities|event.Components,
+	)
 
 	// Get component IDs
 	posID := ecs.ComponentID[Position](&world)
@@ -58,7 +43,7 @@ func main() {
 	rotID := ecs.ComponentID[Rotation](&world)
 
 	// Register a listener function.
-	world.SetListener(&wrapper)
+	world.SetListener(&listener)
 
 	// Create/manipulate/delete entities and observe the listener's output
 	e0 := world.NewEntity(posID)